src/cloudsql_to_gcs/requirements.txt
mysql-connector-python==9.0.0
google-cloud-storage==2.17.0